Initial commit of blog posts
Goal of the repo is to hold work in progress logs of what I've been
doing along with resources that are somewhat useful. Using github to
back it up.

The end goal would likely be to move this to an actual blog, but for now
a collection of files feels adequate enough.